collectionView 역할 분리
